+++ src/middleware/transport.h
@@ -4,6 +4,7 @@
 #include "transport-common.h"
 #include "transport-interthread.h"
 #include "transport-interprocess.h"
+#include "transport-interprocess-zeromq.h"
 #include "transport-intervehicle.h"
 
 #endif
